hrhelpdesk/FAQchatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0

# OpenAI API
openai==1.12.0

# RAG Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
lxml==4.9.3
html5lib==1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# Text processing
pyarabic==0.6.14
deep-translator==1.11.4